+++ gems/smithy/lib/smithy/welds/plugins.rb
@@ -13,6 +13,7 @@ require 'smithy-client/plugins/param_validator'
 require 'smithy-client/plugins/protocol'
 require 'smithy-client/plugins/raise_response_errors'
 require 'smithy-client/plugins/request_compression'
+require 'smithy-client/plugins/resolve_auth'
 require 'smithy-client/plugins/response_target'
 require 'smithy-client/plugins/retry_errors'
 require 'smithy-client/plugins/stub_responses'
@@ -42,6 +43,7 @@ module Smithy
           Smithy::Client::Plugins::Protocol => { require_path: "#{base_path}/protocol" },
           Smithy::Client::Plugins::RaiseResponseErrors => { require_path: "#{base_path}/raise_response_errors" },
           Smithy::Client::Plugins::RequestCompression => { require_path: "#{base_path}/request_compression" },
+          Smithy::Client::Plugins::ResolveAuth => { require_path: "#{base_path}/resolve_auth" },
           Smithy::Client::Plugins::ResponseTarget => { require_path: "#{base_path}/response_target" },
           Smithy::Client::Plugins::RetryErrors => { require_path: "#{base_path}/retry_errors" },
           Smithy::Client::Plugins::StubResponses => { require_path: "#{base_path}/stub_responses" }
